% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_credentials.R
\name{cluster_credentials}
\alias{cluster_credentials}
\title{Get GHAP credentials provided to configure_cluster}
\usage{
cluster_credentials()
}
\description{
Get GHAP credentials provided to configure_cluster
}
