% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_credentials.R
\name{configure_cluster}
\alias{configure_cluster}
\title{Provide cluster credentials for use with run_on_cluster}
\usage{
configure_cluster(json_file = "~/cluster_credentials.json")
}
\arguments{
\item{json_file}{the configuration file}
}
\description{
Provide cluster credentials for use with run_on_cluster
}
