% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{configure_dataset}
\alias{configure_dataset}
\title{configure dataset for use with tlapp}
\usage{
configure_dataset(dataset_path)
}
\arguments{
\item{dataset_path}{path to the datset (for now csv files only)}
}
\description{
configure dataset for use with tlapp
}
