% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghap.R
\name{configure_ghap}
\alias{configure_ghap}
\title{Provide ghap credentials for use with run_on_cluster}
\usage{
configure_ghap(username, password, ip, tlapp_token)
}
\arguments{
\item{username}{your GHAP username}

\item{password}{your GHAP password}

\item{ip}{the ip of your GHAP node}

\item{tlapp_token}{an api token for the tlapp}
}
\description{
Provide ghap credentials for use with run_on_cluster
}
