% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longbow_api.R
\name{force_job_finish}
\alias{force_job_finish}
\title{Longbow Job API}
\usage{
force_job_finish(job_id)
}
\arguments{
\item{job_id}{the job_id of the job being requested. Returned by \code{\link{run_on_longbow}}}
}
\description{
Longbow Job API
}
