% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longbow_api.R
\name{get_job_logs}
\alias{get_job_logs}
\title{Longbow Job API}
\usage{
get_job_logs(job_id)
}
\arguments{
\item{job_id}{the job_id of the job being requested. Returned by \code{\link{run_on_longbow}}}
}
\description{
Longbow Job API
}
