% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_helpers.R
\name{inputs_json_from_rmd}
\alias{inputs_json_from_rmd}
\title{Generate sample inputs.json object using defaults specified in Rmd file}
\usage{
inputs_json_from_rmd(rmd_filename, save_as = NULL)
}
\arguments{
\item{rmd_filename}{the analysis template to use}

\item{save_as}{the filename to save the resulting inputs.json to}
}
\value{
the raw json text
}
\description{
can be used to create an inputs.json file that can then be modified with other paramter values
}
