% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_helpers.R
\name{params_from_rmd}
\alias{params_from_rmd}
\title{Generate sample params object using defaults specified in Rmd file}
\usage{
params_from_rmd(rmd_filename, save_as = NULL)
}
\arguments{
\item{rmd_filename}{the analysis template to use}

\item{save_as}{the json filename to save the resulting params to}
}
\value{
the params object
}
\description{
Can also be used to create an inputs.json file that can then be modified with other paramter values
}
