% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longbow_api.R
\name{publish_template}
\alias{publish_template}
\title{Publish template to longbow UI}
\usage{
publish_template(rmd_filename, open_result = TRUE)
}
\arguments{
\item{rmd_filename}{the template to publish}

\item{open_result}{if \code{TRUE}, opens compiled report in browser}
}
\description{
Publish template to longbow UI
}
