% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longbow_api.R, R/run_analysis.R
\name{run_on_longbow}
\alias{run_on_longbow}
\alias{run_internal}
\alias{run_locally}
\title{Run template with parameters as specified and generate markdown output}
\usage{
run_on_longbow(rmd_filename, params_filename, open_result = TRUE)

run_internal(rmd_filename, params_filename, output_directory = tempdir())

run_locally(rmd_filename, params_filename, output_directory = tempdir(),
  open_result = TRUE)
}
\arguments{
\item{rmd_filename}{the script to template}

\item{params_filename}{a json file specifying the parameters to use}

\item{open_result}{if \code{TRUE}, opens compiled report in browser}

\item{output_directory}{a folder to work in}
}
\description{
Run template with parameters as specified and generate markdown output
}
