% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_locally}
\alias{run_locally}
\title{Run script with parameters as specified and generate markdown output}
\usage{
run_locally(rmd_filename, params_filename, output_directory = tempdir())
}
\arguments{
\item{rmd_filename}{the script to run}

\item{params_filename}{a json file specifying the parameters to use}

\item{output_directory}{a folder to work in}
}
\description{
Run script with parameters as specified and generate markdown output
}
