% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_helpers.R
\name{get_tl_data}
\alias{get_tl_data}
\alias{get_tl_nodes}
\alias{get_tl_params}
\title{Script helpers}
\usage{
get_tl_data(params_object = NULL)

get_tl_nodes(params_object = NULL)

get_tl_params(params_object = NULL)
}
\arguments{
\item{params_object}{the object to use, defaults to the global \code{params} object}
}
\description{
Get relevant objects from script parameters
}
