% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptParams.R
\docType{data}
\name{ScriptParams}
\alias{ScriptParams}
\title{Main TLtools class
todo: write docs}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScriptParams
}
\description{
Main TLtools class
todo: write docs
}
\keyword{datasets}
