% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghap.R
\name{configure_ghap}
\alias{configure_ghap}
\title{Provide ghap credentials for use with run_on_cluster}
\usage{
configure_ghap(json_file = "~/ghap.json")
}
\arguments{
\item{json_file}{the configuration file}
}
\description{
Provide ghap credentials for use with run_on_cluster
}
