% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_internal}
\alias{run_internal}
\alias{run_locally}
\alias{run_on_cluster}
\title{Run script with parameters as specified and generate markdown output}
\usage{
run_internal(rmd_filename, params_filename, output_directory = tempdir())

run_locally(rmd_filename, params_filename, open_result = TRUE)

run_on_cluster(rmd_filename, params_filename, open_result = TRUE)
}
\arguments{
\item{rmd_filename}{the script to run}

\item{params_filename}{a json file specifying the parameters to use}

\item{output_directory}{a folder to work in}

\item{open_result}{Open compiled report in new window (Currently OSX only)}
}
\description{
Run script with parameters as specified and generate markdown output
}
